[feat] entity (#21)
