+++ requirements.txt
@@ -4,4 +4,7 @@ python-dotenv==1.0.0
 marshmallow==3.20.1
 pytest==7.4.3
 httpx==0.25.2
+sqlalchemy==2.0.23
+pymysql==1.1.0
+cryptography==41.0.5
 pre-commit
